api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.3
